+++ inst/NEWS.md
@@ -1,18 +1,24 @@
 __MEDseq: Mixtures of Exponential-Distance Models with Covariates__   
 ===================================================================
 
-### Improvements, Bug Fixes & Miscellaneous Edits
+## MEDseq v1.4.0 - (_11<sup>th</sup> release [minor update]: 2022-12-20_)
+### New Features & Improvements
 * Function `seqdef` added as an exact copy of `TraMineR::seqdef`, to enable experienced  
   users of `MEDseq` & `TraMineR` to use the former without needing to explicitly load the latter.
 * `MEDseq_clustnames` gains the arg. `weighted=FALSE` for use when `size=TRUE`:  
   this is now respected by the `weighted` arg. to `plot.MEDseq` where relevant.
+* New function `dist_freqwH` added for calculating pairwise dissimilarity matrix associated with  
+  `wKModes(..., freq.weighted=TRUE)` for subsequent use (e.g. silhouettes).
 * The `plot.MEDseq` function's `type` arg. gains the option `"dH"`,  
   provided version `2.2-4` or later of the `TraMineR` package is installed.
 * `plot.MEDseq` also gains the `"similarity"` option for its `type` argument.
+* New function `MEDseq_AvePP` added.
+
+### Bug Fixes & Miscellaneous Edits
 * `wKModes` now also returns `x$tot.withindiff` (i.e. `sum(x$withindiff)`).
 * Minor speed-ups to `wKModes` when `freq.weighted=TRUE`.
 * Minor cosmetic changes to `type="dbsvals"` & `type="aswvals"` in `plot.MEDseq`.
-* New function `MEDseq_AvePP` added.
+* Minor speed-ups to `plot.MEDseq` related to its `seriated` arg. in `G=1` settings.
 * Fixed rare bugs in tie-breaking for modal sequence estimate in `MEDseq_fit` & `wKModes`.
 * Fixed documentation typos.